+++ source/Core/Email.php
@@ -280,7 +280,7 @@ class Email extends \PHPMailer
         //enabling exception handling in phpMailer class
         parent::__construct(true);
 
-        $myConfig = $this->getConfig();
+        $myConfig = Registry::getConfig();
 
         $this->_setMailerPluginDir();
         $this->setSmtp();
@@ -320,18 +320,6 @@ class Email extends \PHPMailer
     }
 
     /**
-     * oxConfig instance getter
-     *
-     * @deprecated since v6.4.0 (2018-10-15); This method will be removed completely. Use \OxidEsales\Eshop\Core\Registry::getConfig().
-     *
-     * @return \OxidEsales\Eshop\Core\Config
-     */
-    public function getConfig()
-    {
-        return \OxidEsales\Eshop\Core\Registry::getConfig();
-    }
-
-    /**
      * Smarty instance getter, assigns this oxEmail instance to "oEmailView" variable
      *
      * @return smarty
@@ -362,7 +350,7 @@ class Email extends \PHPMailer
             return false;
         }
 
-        $myConfig = $this->getConfig();
+        $myConfig = Registry::getConfig();
         $this->setCharSet();
 
         if ($this->_getUseInlineImages()) {
@@ -440,7 +428,7 @@ class Email extends \PHPMailer
      */
     public function setSmtp($shop = null)
     {
-        $myConfig = $this->getConfig();
+        $myConfig = Registry::getConfig();
         $shop = ($shop) ? $shop : $this->_getShop();
 
         $smtpUrl = $this->_setSmtpProtocol($shop->oxshops__oxsmtp->value);
@@ -555,7 +543,7 @@ class Email extends \PHPMailer
      */
     public function sendOrderEmailToOwner($order, $subject = null)
     {
-        $config = $this->getConfig();
+        $config = Registry::getConfig();
 
         $shop = $this->_getShop();
 
@@ -823,7 +811,7 @@ class Email extends \PHPMailer
      */
     protected function _getNewsSubsLink($id, $confirmCode = null)
     {
-        $myConfig = $this->getConfig();
+        $myConfig = Registry::getConfig();
         $actShopLang = $myConfig->getActiveShop()->getLanguage();
 
         $url = $myConfig->getShopHomeUrl() . 'cl=newsletter&amp;fnc=addme&amp;uid=' . $id;
@@ -881,7 +869,7 @@ class Email extends \PHPMailer
      */
     public function sendSuggestMail($params, $product)
     {
-        $myConfig = $this->getConfig();
+        $myConfig = Registry::getConfig();
 
         //sets language of shop
         $currLang = $myConfig->getActiveShop()->getLanguage();
@@ -938,7 +926,7 @@ class Email extends \PHPMailer
      */
     public function sendInviteMail($params)
     {
-        $myConfig = $this->getConfig();
+        $myConfig = Registry::getConfig();
 
         //sets language of shop
         $currLang = $myConfig->getActiveShop()->getLanguage();
@@ -1002,7 +990,7 @@ class Email extends \PHPMailer
      */
     public function sendSendedNowMail($order, $subject = null)
     {
-        $myConfig = $this->getConfig();
+        $myConfig = Registry::getConfig();
 
         $orderLang = (int) (isset($order->oxorder__oxlang->value) ? $order->oxorder__oxlang->value : 0);
 
@@ -1070,7 +1058,7 @@ class Email extends \PHPMailer
      */
     public function sendDownloadLinksMail($order, $subject = null)
     {
-        $myConfig = $this->getConfig();
+        $myConfig = Registry::getConfig();
 
         $orderLang = (int) (isset($order->oxorder__oxlang->value) ? $order->oxorder__oxlang->value : 0);
 
@@ -1245,7 +1233,7 @@ class Email extends \PHPMailer
 
             $this->setRecipient($shop->oxshops__oxowneremail->value, $shop->oxshops__oxname->getRawValue());
             $this->setFrom($shop->oxshops__oxowneremail->value, $shop->oxshops__oxname->getRawValue());
-            $this->setBody($smarty->fetch($this->getConfig()->getTemplatePath($this->_sReminderMailTemplate, false)));
+            $this->setBody($smarty->fetch(Registry::getConfig()->getTemplatePath($this->_sReminderMailTemplate, false)));
             $this->setAltBody("");
             $this->setSubject(($subject !== null) ? $subject : $lang->translateString('STOCK_LOW'));
 
@@ -1970,11 +1958,11 @@ class Email extends \PHPMailer
             if (isset($this->_oShop)) {
                 return $this->_oShop;
             } else {
-                return $this->_oShop = $this->getConfig()->getActiveShop();
+                return $this->_oShop = Registry::getConfig()->getActiveShop();
             }
         }
 
-        $myConfig = $this->getConfig();
+        $myConfig = Registry::getConfig();
 
         $shop = oxNew(\OxidEsales\Eshop\Application\Model\Shop::class);
         if ($shopId !== null) {
@@ -2117,7 +2105,7 @@ class Email extends \PHPMailer
      */
     public function getViewConfig()
     {
-        return $this->getConfig()->getActiveView()->getViewConfig();
+        return Registry::getConfig()->getActiveView()->getViewConfig();
     }
 
     /**
@@ -2127,7 +2115,7 @@ class Email extends \PHPMailer
      */
     public function getView()
     {
-        return $this->getConfig()->getActiveView();
+        return Registry::getConfig()->getActiveView();
     }
 
     /**
@@ -2225,7 +2213,7 @@ class Email extends \PHPMailer
      */
     private function _clearSidFromBody($altBody)
     {
-        return \OxidEsales\Eshop\Core\Str::getStr()->preg_replace('/(\?|&(amp;)?)(force_)?(admin_)?sid=[A-Z0-9\.]+/i', '\1shp=' . $this->getConfig()->getShopId(), $altBody);
+        return \OxidEsales\Eshop\Core\Str::getStr()->preg_replace('/(\?|&(amp;)?)(force_)?(admin_)?sid=[A-Z0-9\.]+/i', '\1shp=' . Registry::getConfig()->getShopId(), $altBody);
     }
 
     /**
@@ -2243,7 +2231,7 @@ class Email extends \PHPMailer
      */
     private function isDebugModeEnabled()
     {
-        return $this->getConfig()->getConfigParam('iDebug') != 0;
+        return Registry::getConfig()->getConfigParam('iDebug') != 0;
     }
 
     /**
@@ -2260,7 +2248,7 @@ class Email extends \PHPMailer
           WHERE `OXACTIVE` = 1 
           AND `OXUSERNAME` = ? 
           AND `OXPASSWORD` != ''";
-        if ($this->getConfig()->getConfigParam('blMallUsers')) {
+        if (Registry::getConfig()->getConfigParam('blMallUsers')) {
             $select .= "ORDER BY OXSHOPID = ? DESC";
         } else {
             $select .= "AND OXSHOPID = ?";